Accession Class

An S4 Class to represent a breeDBase accession

Details

The `accession_name` and `species_name` are required, all other fields are optional.

Slots

accession_name

Accession Name (must be unique)

species_name

Species Name (must exist in the database)

population_name

(optional) a population is a grouping of accessions. if the population already exists in the database, the accession will be added into it, otherwise, the new population will be created.

organization_names

(optional) the names of the organizations which use this accession.

synonyms

(optional) an accession can be known by many names including local popular names. a synonym name can be used instead of the accession_name throughout the database; because of this, synonyms must themselves be unique.

location_codes

(optional) location codes for the accession.

ploidy_levels

(optional) a number indicating the ploidy (e.g. 2 for diploid, 3 for triploid), numeric.

genome_structures

(optional) genome structures for accession which take into account ploidy and ancestral genome info e.g. AAA, AB.

varietys

(optional) variety can be defined as a group of individuals or plants having similar traits that can be reproduced "true to type" from generation to generation.

donors

(optional) the accession_name of the donor accession.

donor_institutes

(optional) the institute of the donor accession.

donor_PUIs

(optional) the permanent unique identifier of the donor accession.

country_of_origins

(optional) the country of origin.

states

(optional) the state of origin.

institute_codes

(optional) the institute code of origin.

institute_names

(optional) the institute name of origin.

biological_status_of_accession_codes

(optional) code indicating the state of accession.

notes

(optional) free text for notes.

accession_numbers

(optional) accession number for accession from germplasm bank.

PUIs

(optional) permanent unique identifier of the accession.

seed_sources

(optional) origin of seed source.

type_of_germplasm_storage_codes

(optional) code indicating the type of germplasm storage.

acquisition_dates

(optional) dates of acquisition YYYYMMDD for accession.

transgenic

(optional) indicates if accession is transgenic. Please indicate 1 if transgenic and empty if otherwise.

introgression_parent

(optional) if the accession you are adding has an introgression that originated from one of the parents, you can specify the parent here

introgression_backcross_parent

(optional) the backcross parent for introducing an introgression into the accession being added

introgression_map_version

(optional) the map version for identifying the start and stop position of the introgression. e.g. AGPv2

introgression_chromosome

(optional) the chromosome number that the introgression is on.

introgression_start_position_bp

(optional) the start position of the introgression in base pairs

introgression_end_position_bp

(optional) the end position of the introgression in base pairs

purdy_pedigree

(optional, experimental) the purdy pedigree string

filial_generation

(optional, experimental) the filial generation number