Create Plot

Create a phenotyping trial Plot containing all of the provided properties. Plot Name, Accession Name, Plot Number and Block Number are required. All optional properties can be added as a named value in the properties list.

Plot(
  plot_name = NULL,
  accession_name = NULL,
  plot_number = NULL,
  block_number = NULL,
  properties = list()
)

Arguments

plot_name

The unique name for the plot (must be unique across entire database. this is often a concatenation of the trial name, the accession name, and the plot number.)

accession_name

The name of the accession being tested in the plot (must exist in the database)

plot_number

The sequential number for the plot in the field (e.g. 1001, 1002, 2001, 2002). these numbers should be unique for the trial.

block_number

A design parameter indicating which block the plot is in

properties

(optional) Additional plot properties (as a named list)

Value

Plot

Details

See Class Plot for all optional plot properties

See also

Examples

# Create plots with just the required parameters
plot1 <- Plot("FARM-2019-UNH_PLOT1", "SL18-UCONN-S131", 1, 1)
plot2 <- Plot("FARM-2019-UNH_PLOT2", "SL18-UCONN-S31", 2, 1)

# Create plots with optional parameters
plot3 <- Plot("FARM-2019-UNH_PLOT3", "SL18-UCONN-S105", 3, 1, list(row_number = 1, col_number = 3))
plot4 <- Plot("FARM-2019-UNH_PLOT4", "SL18-UCONN-S146", 4, 1, list(row_number = 1, col_number = 4))

# Create a plot with an applied treatment
plot5 <- Plot(
     "FARM-2019-UNH_PLOT5", 
     "SL18-UCONN-S110", 
     5, 
     1, 
     list(
         row_number = 1, 
         col_number = 5, 
         treatments = list(
             inoculated = TRUE
         )
     )
)