Plot Class

An S4 Class to represent a breeDBase phenotyping trial plot

Details

The `plot_name`, `accession_name`, `plot_number`, and `block_number` are required, all other fields are optional

Slots

plot_name

The unique name for the plot (must be unique across entire database. this is often a concatenation of the trial name, the accession name, and the plot number.)

accession_name

The name of the accession being tested in the plot (must exist in the database)

plot_number

The sequential number for the plot in the field (e.g. 1001, 1002, 2001, 2002). these numbers should be unique for the trial.

block_number

A design parameter indicating which block the plot is in

is_a_control

(optional) TRUE if the plot is a control. Generally, you will have accessions that are controls, so you should indicate the plots that that accession is in as a control.

rep_number

(optional) replicate number

range_number

(optional) range number, often synonymous with col_number

row_number

(optional) row number. if the field is a grid, this represents the y coordinate, required for field map generation. the top left plot should be row 1, column 1

col_number

(optional) column number. if the field is a grid, this represents the x coordinate. sometimes called range_number, required for field map generation. the top left plot shuold be row 1, column 1

seedlot_name

(optional) the seedlot from where the planted seed originated. must exist in the database

num_seed_per_plot

(optional) number seeds per plot. seed is transferred from seedlot mentioned in seedlot_name.

weight_gram_seed_per_plot

(optional) weight in gram of seeds in plot. seed is transferred from seedlot mentioned in seedlot name.

treatments

(optional) A named list where the name is the name of a treatment (e.g. inoculated, drought, etc) and the value should be TRUE if the treatment was applied to the plot or FALSE if not.

See also