Create Trial

Create a phenotyping Trial containing all of the provided properties. Trial Name, Breeding Program, Location, Year, Design Type and Description are required. All optional properties can be added as a named value in the properties list.

Trial(
  trial_name = NULL,
  breeding_program = NULL,
  location = NULL,
  year = NULL,
  design_type = NULL,
  description = NULL,
  properties = list()
)

Arguments

trial_name
A name for the trial (must be unique across the entire database)

breeding_program
The name of the breeding program that managed the trial (must exist in database)

location
The name of the location where the trial was held (must exist in database)

year
The year the trial was held

design_type
The shorthand for the design type (CRD, RCBD, Alpha, Augmented, MAD, Westcott, Lattice)

description
Additional text with any other relevant information about the trial

properties
(optional) Additional trial properties (as a named list)

Value

Trial

Details

See Class Trial for all optional trial properties

Use the getTrialDesignTypes function to get a list of supported Trial design types and the getTrialTypes function to get a list of the supported Trial types.

See also

Examples

# Create a Trial with just the required parameters
trial1 <- Trial("UMOPN_2019_Madison", "University of Wisconsin", "Madison, WI", 2019, "RCBD", "UMOPN Nursery Trial")

# Create a Trial with optional parameters
opts <- list(planting_date = "2019-04-25", harvest_date = "2019-10-05", trial_type = "phenotyping_trial")
trial2 <- Trial("UMOPN_2019_Arlington", "University of Wisconsin", "Arlington, WI", 2019, "RCBD", "UMOPN Nursery Trial", opts)